feat: authorization implemented

+++ notes-app-back-end-plugin/src/api/notes/handler.js
@@ -16,12 +16,15 @@ class NotesHandler {
 
   async postNoteHandler(request, h) {
     try {
-      // bisa access method karena udah di-register
       this._validator.validateNotePayload(request.payload);
-      const { title = 'untitled', tags, body } = request.payload;
-
-      // return id from addNote
-      const noteId = await this._service.addNote({ title, tags, body });
+      const { title = 'untitled', body, tags } = request.payload;
+      const { id: credentialId } = request.auth.credentials;
+      const noteId = await this._service.addNote({
+        title,
+        body,
+        tags,
+        owner: credentialId,
+      });
 
       const response = h.response({
         status: 'success',
@@ -38,25 +41,24 @@ class NotesHandler {
           status: 'fail',
           message: error.message,
         });
-
         response.code(error.statusCode);
         return response;
       }
 
-      console.error(error);
+      // Server ERROR!
       const response = h.response({
         status: 'error',
-        message: 'Maaf terjadi kegagalan di server kami',
+        message: 'Maaf, terjadi kegagalan pada server kami.',
       });
-
       response.code(500);
+      console.error(error);
       return response;
     }
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
-
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -68,7 +70,11 @@ class NotesHandler {
   async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
       const note = await this._service.getNoteById(id);
+
       return {
         status: 'success',
         data: {
@@ -81,53 +87,50 @@ class NotesHandler {
           status: 'fail',
           message: error.message,
         });
-
         response.code(error.statusCode);
         return response;
       }
 
-      console.error(error);
+      // Server ERROR!
       const response = h.response({
         status: 'error',
-        message: 'Maaf terjadi kegagalan di server kami',
+        message: 'Maaf, terjadi kegagalan pada server kami.',
       });
-
       response.code(500);
+      console.error(error);
       return response;
     }
   }
 
   async putNoteByIdHandler(request, h) {
     try {
-      // bisa access method karena udah di-register
       this._validator.validateNotePayload(request.payload);
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+      await this._service.verifyNoteOwner(id, credentialId);
       await this._service.editNoteById(id, request.payload);
 
-      const response = h.response({
+      return {
         status: 'success',
         message: 'Catatan berhasil diperbarui',
-      });
-      response.code(200);
-      return response;
+      };
     } catch (error) {
       if (error instanceof ClientError) {
         const response = h.response({
           status: 'fail',
           message: error.message,
         });
-
         response.code(error.statusCode);
         return response;
       }
 
-      console.error(error);
+      // Server ERROR!
       const response = h.response({
         status: 'error',
-        message: 'Maaf terjadi kegagalan di server kami',
+        message: 'Maaf, terjadi kegagalan pada server kami.',
       });
-
       response.code(500);
+      console.error(error);
       return response;
     }
   }
@@ -135,32 +138,31 @@ class NotesHandler {
   async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+      await this._service.verifyNoteOwner(id, credentialId);
       await this._service.deleteNoteById(id);
 
-      const response = h.response({
+      return {
         status: 'success',
         message: 'Catatan berhasil dihapus',
-      });
-      response.code(200);
-      return response;
+      };
     } catch (error) {
       if (error instanceof ClientError) {
         const response = h.response({
           status: 'fail',
           message: error.message,
         });
-
         response.code(error.statusCode);
         return response;
       }
 
-      console.error(error);
+      // Server ERROR!
       const response = h.response({
         status: 'error',
-        message: 'Maaf terjadi kegagalan di server kami',
+        message: 'Maaf, terjadi kegagalan pada server kami.',
       });
-
       response.code(500);
+      console.error(error);
       return response;
     }
   }
